pyproject.toml
[project]
name = "inv-classifier"
version = "0.1.0"
description = "This project an automation system that help to digitalize and classify invoice using ai"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
